moved stageProps back into stage. use Stage factory functions in tests and removed mock
