+++ src/Stage.js
@@ -2,14 +2,61 @@ import React, { useState, useEffect, useLayoutEffect, useRef } from "react";
 import PropTypes from "prop-types";
 import * as PIXI from "pixi.js";
 import { AppProvider } from "./AppProvider";
-import { getCanvasProps, getDisplayObjectProps, propTypes } from "./stageProps";
 import { DEFAULT_PROPS, EVENT_PROPS } from "./props";
 import { applyProps } from "./ReactPixiFiber";
 import { render, unmount } from "./render";
 import { usePreviousProps, usePixiApp } from "./hooks";
-import { including } from "./utils";
+import { filterByKey, including } from "./utils";
 
 export let appTestHook = null;
+
+export function validateCanvas(props, propName, componentName) {
+  // Let's assume that element is canvas if the element is Element and implements getContext
+  const element = props[propName];
+  if (typeof element === "undefined") {
+    return;
+  }
+
+  const isCanvas = element instanceof Element && typeof element.getContext === "function";
+  if (!isCanvas) {
+    const propType = typeof element;
+    return new Error(
+      `Invalid prop '${propName}' of type '${propType}' supplied to '${componentName}', expected '<canvas> Element'.`
+    );
+  }
+}
+
+export const propTypes = {
+  options: PropTypes.shape({
+    antialias: PropTypes.bool,
+    autoStart: PropTypes.bool,
+    backgroundColor: PropTypes.number,
+    clearBeforeRender: PropTypes.bool,
+    forceCanvas: PropTypes.bool,
+    forceFXAA: PropTypes.bool,
+    height: PropTypes.number,
+    legacy: PropTypes.bool,
+    powerPreference: PropTypes.string,
+    preserveDrawingBuffer: PropTypes.bool,
+    resolution: PropTypes.number,
+    roundPixels: PropTypes.bool,
+    sharedLoader: PropTypes.bool,
+    sharedTicker: PropTypes.bool,
+    transparent: PropTypes.bool,
+    view: validateCanvas,
+    width: PropTypes.number,
+  }),
+  children: PropTypes.node,
+  height: PropTypes.number,
+  width: PropTypes.number,
+};
+
+export const includingDisplayObjectProps = including(Object.keys(DEFAULT_PROPS).concat(EVENT_PROPS));
+export const includingStageProps = including(Object.keys(propTypes));
+export const includingCanvasProps = key => !includingDisplayObjectProps(key) && !includingStageProps(key);
+
+export const getCanvasProps = props => filterByKey(props, includingCanvasProps);
+export const getDisplayObjectProps = props => filterByKey(props, includingDisplayObjectProps);
 
 const applyUpdate = (app, props, instance) => {
   const provider = <AppProvider app={app}>{props.children}</AppProvider>;
@@ -81,8 +128,8 @@ export function createStageClass() {
     componentDidUpdate(prevProps) {
       const { children, height, options, width } = this.props;
       const { options: prevOptions } = prevProps;
-      resizeRenderer(this._app, prevProps, this.props);
       applyUpdate(this._app, this.props, this);
+      resizeRenderer(this._app, prevProps, this.props);
     }
 
     componentWillUnmount() {
