+++ src/index.js
@@ -3,11 +3,30 @@ import { AppContext, AppProvider, withApp } from "./AppProvider";
 import Stage from "./Stage";
 import { TYPES } from "./types";
 import { render } from "./render";
+import {
+  usePixi,
+  usePixiTicker,
+  usePreviousProps,
+  usePixiApp,
+} from "./hooks";
 import { applyProps, unstable_batchedUpdates } from "./ReactPixiFiber";
 
 /* Public API */
 
-export { AppContext, AppProvider, CustomPIXIComponent, Stage, applyProps, render, withApp, unstable_batchedUpdates };
+export {
+  AppContext,
+  AppProvider,
+  CustomPIXIComponent,
+  Stage,
+  applyProps,
+  render,
+  withApp,
+  unstable_batchedUpdates,
+  usePixi,
+  usePixiTicker,
+  usePreviousProps,
+  usePixiApp,
+};
 
 export const BitmapText = TYPES.BITMAP_TEXT;
 export const Container = TYPES.CONTAINER;
